requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
groq>=0.4.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
pypdf>=3.17.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
